requirements.txt
requests==2.13.0
PyYAML==3.12


